+++ packages/web-client/src/pages/CharacterSelect.tsx
@@ -1,11 +1,10 @@
 import React, { useEffect, useState } from "react";
-import { useNavigate } from "react-router-dom";
 import WalletSelector from "../components/wallet/WalletSelector";
-import { IoMdClose } from "react-icons/io";
 import useWalletConnection from "../hooks/useWalletConnection";
 import supabase from "../config";
 import AvatarCard from "../components/layout/AvatarCard";
 import { Avatar } from "../types/Avatar";
+import UserAccount from "../components/layout/UserAccount";
 
 const CharacterSelect: React.FC = () => {
   // Initialize with empty array instead of undefined
@@ -48,27 +47,22 @@ const CharacterSelect: React.FC = () => {
   const [selectedCharacter, setSelectedCharacter] = useState<number | null>(
     null
   );
-  const navigate = useNavigate();
   const { isConnected, activeAddress } = useWalletConnection();
 
-   const [modal, setModal] = useState(false);
+  const [modal, setModal] = useState(false);
 
-   const toggleModal = () => {
+  const toggleModal = () => {
      setModal(!modal);
-   };
+  };
 
-   if (modal) {
+  if (modal) {
      document.body.classList.add("overflow-y-hidden");
-   } else {
+  } else {
      document.body.classList.remove("overflow-y-hidden");
-   }
+  }
 
   const handleSelect = (id: number) => {
     setSelectedCharacter(id);
-  };
-
-  const handleStartGame = () => {
-    navigate("/play-game");
   };
 
   // If not connected, don't render the page content
@@ -130,38 +124,7 @@ const CharacterSelect: React.FC = () => {
       )}
 
       {modal && (
-        <div className="w-screen h-screen top-0 left-0 right-0 bottom-0 fixed">
-          <div className="w-screen h-screen top-0 left-0 right-0 bottom-0 fixed bg-black/50">
-            <div
-              className="absolute w-xl h-auto top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 
-              bg-black py-3 px-5 rounded-md max-w-2xl min-w-xs border-5 border-white"
-            >
-              <form className="flex flex-col items-center">
-                <h2 className="font-['Alexandria'] font-bold text-2xl pb-2">
-                  Username
-                </h2>
-                <IoMdClose
-                  size={30}
-                  onClick={toggleModal}
-                  className="absolute cursor-pointer right-5"
-                />
-                <input
-                  type="text"
-                  id="username"
-                  placeholder="(e.g. John Doe)"
-                  className="font-['Alexandria'] px-2 pt-2 mb-3 text-xl focus:outline-none w-lg border-b-2 border-white"
-                />
-                <button
-                  type="submit"
-                  className="cursor-pointer bg-green-700 font-['Alexandria'] 
-                   text-xl font-bold hover:opacity-75 w-xs align-center rounded-lg my-3 p-2 border-b-5 border-green-900 "
-                 onClick={handleStartGame}>
-                  PLAY
-                </button>
-              </form>
-            </div>
-          </div>
-        </div>
+        <UserAccount />
       )}
     </div>
   );
